Added execution max idle parameter
Execution will be aborted if it was idle for more then the max idle value (by default 30 seconds).
If execution is aborted and messages about this exeuction is reached from other shards,
then it will be ignored, which means that eventually the execution will be timedout on all the shards.