Fix #619, use `getkeys-api` flag to allow Redis to catch MOVED error. (#673)
* Fix #619, use `getkeys-api` flag to allow Redis to catch MOVED error.

Hooking a command works using a command filter that direct a command to
another internal command that runs the hook code. So for example, if the
use sends `get x`, Redis will actually see `rg.trigger get x`. Redis
do not know that on `rg.trigger get x`, `x` is actually a key and it
will not send MOVED error on cluster.

The fix, use `getkeys-api` to allow to tell Redis dynamically where keys
are. When `rg.trigger get x` is invoke, Redis will ask RedisGears to tell
it where are the keys in the command and gears will dynamically calculate
the keys base on the original command that was executed.

Tests was added to verify the fix.

* Fix tests