Fix globals dictionary not set correctly after deserialization. (#676)
The problem
---

After deserialization, the deserialized function are created with its own
globals dictionary. We want all the functions on the same execution (and even different
execution that was generated by the same `RG.PYEXECUTE` command) to share the same globals
dictionary. The current solution is to change the function global dictionary to be the
global dictionary of the `PY.EXECUTE` session that created the execution.

This solution only changes the top most function, but it does not change function that
is used by it, so the following example will break:

```
g = 1
def f1():
    global g
    g = g + 1
    print(id(globals()))
    return g

def f(x):
    global g
    g = g + 1
    print(id(globals()))
    return f1()

GB('CommandReader').map(f).register(trigger='test', convertToStr=False)
GB('CommandReader').map(f).register(trigger='test1', convertToStr=False)
```

Running `RG.TRIGGER test` and `RG.TRIGGER test1` will give the following:

```
> RG.TRIGGER test
3
> RG.TRIGGER test1
5
```

But after RDB load we will get the following result (that proves that `f1` was
created with a different global dictionary):

```
> RG.TRIGGER test
3
> RG.TRIGGER test1
4
```

The solution
---

Modify `cloudpickle` to look for a global variable, `requested_base_globals`. If set,
use this dictionary as the global dictionary of all the created functions. RedisGears
will make sure to set `requested_base_globals` to the correct value before deserializing
a function. In addition, changed `cloudpickle` to update the global dictionary only with
the missing fields and not override existing fields. After this change, the above example
works as expected.

Test was added to verify the fix.

(cherry picked from commit fe89d09a8c9192cddda486afb290973953b5ac9c)