Extract more information on test failure
Extract the following information of test failure:
1. RG.DUMPREGISTRATIONS
2. RG.DUMPEXECUTIONS
3. info everything

The information will be written to a file on the log directory.
File name will be the same as the Redis log file name with a suffix
contains the test name. This will promise that the collected informatio
will be uploaded to the CI and we will be able to analize it.