See #724, add support for registrations pause and unpause.
The PR introduce 2 new commands:

* `RG.PAUSEREGISTRATIONS [ABORTPENDINGS] <ID1> [<ID2> ...]`
* `RG.UNPAUSEREGISTRATIONS <ID1> [<ID2> ...]`

The new commands allows to pause/unpause registration from triggering new
executions. Currently, only stream registrations support pausing/unpausing.
An attempt to pause/unpause a registration that do not use the stream reader
will raise an error.

pause/unpause a registration that already pause/unpause is consider as no op
and will keep the state exactly as it is (without any errors).

pause/unpause is considered an atomic operation, all the registrations that
was given will be pause/unpause together and if one failed, the entire operation
is aborted (atomicity is promised on the shard level and not on the cluster level).

Unpausing a registration will restart processing data from the stream, **if the stream
is not set to trim messages (using `trimStream` option) then all the element will
potentialy be processed again**.

This PR also fix #722, added `trimStream`, `onFailedPolicy` and `retryInterval`
to `RG.DUMPREGISTRATIONS`.