Fix #792, trigger execution infinitely when trimming is disabled. (#798)
* Fix #792, trigger execution infinitly when trimming is disabled.

When trimming is disabled, if the registration was registered
on a strim such that gears consumer group are not point to the begining
of the stream, RedisGears would enter a state where it triggers
execution infinitly. Such situation can happened when upgrade a
registrations (that has trimming disable) or when loading from RDB
(again when trimming disabled).

This happened because RedisGears decides to trigger executions base
on the amount of pending data which is calculated by checking the stream lenght.
This check is wrong when trimming is disabled because the stream lenght
is not a good indication to the amount of pending data.

To solve this issue, RedisGears will use a different way to decide if there
is more data to processes. If, in the current batch, there was no data to processes
and no new data was added to the stream durring the batch execution, we know there is
no more data to processes and we avoid trigger another execution.

* Set pending messages to 0 when reaching the end of the stream.

* review fixes

* review fixes