Introduce JS api versioning.
This introduces a way to version the JS api.

The JS api versioning is mandatory for all user modules.
It can be specified via the prologue property "api_version", for
example:

    #!js api_version=1.0 name=lib

This version will be checked and a correct implementation (if supported
and found) will be chosen and provided to the module.