Split the backend into initialised and uninitialised.
This effectively provides a much safer code compared to what has been
before. Before the code implied knowledge and certain code paths to be
taken in order to use the backend functions correctly: certain functions
should have only be used after initialisation and there were no checks
done for that. It was very easy for a developer to forget that fact as
it is unreliable (how can one know which functions can only be called on
an initialised backend object and which can be called on a
non-yet-initialised one).

This commit makes the backend objects have two stages: uninitialised and
initialised. The uninitialised backend only provides a way to obtain its
name and initialise it, giving a reliable way of knowing what can be
called and not. The initialised backend is proven to be initialised by
the Rust's type safety mechanisms and can be safely used everywhere.