Refactoring due to redismodule-rs changes.
Till now, RedisGears-2 was working with a fork of redismodule-rs
that introduce new API's that RedisGears required to work correctly. Those API are on their way to be merge upstream to the master branch
and soon to be on an officail redismodule-rs version. As a result, RedisGears first intrests is to move the redismodule-r
master branch (and an official version as soon as it will be released). The main issue is that some of the API's were implemented different compare to the implementation on the fork that RedisGears was use to work with.

This PR makes the required changes to the RedisGears code to be able to work with redismodule-rs master branch. Though there is a lot of changes, **all the changes are technical and should introduce no behaviure changes what so ever**. As a prove, we can see that the tests changes are minimal and mostly involve error message rephrase.

Changes breakdown:
1. Move to the new configuration API. As a result:
   1.  V8 plugin path is given as `v8-plugin-path` configuration option.
   2.  RedisGears config commands were changed to `RG.CONFIGGET` and `RG.CONFIGSET`
2. Make use or `RedisGILGuard` to protect global object (notice that not all global object were change and we should continue this on a follow up PR).
3. Move support minimal Redis version of 7.2RC1 because of recent fixes that was made on ACL validation when used with `RM_Call`.
4. ACL User was changed to be `RedisString` all over the code. Along side the use of `safe_clone` function.
5. Detached context is used to log messages.
6. Some callbacks were modifify to recieve `&Context` which indicate the Redis GIL is lock and global objects can be access safely.
7. **Avoid a log of unsafe code** (this still an ongoing work and we should aim for zero unsafe blocks on RedisGears code).
8. API between RedisGears core and V8 plugin was modify to correctly reflect lifetime restrictions.
9. Global allocator on V8 plugin was moved, it is now provided by the fact that V8 plugin also put `redismodule-rs` as dependency.

Related work that will be done on future PR's:

1. Get rid of all RefCell, each object should be protected either by mutex or by `RedisGILGuard`. We should also consider introducing a `V8GILGuard` that indicate that an object is protected by the V8 lock.
2. Get rid of all unsafe code