Fix flaky tests.
Redis do not update the client `woff` (Last write global replication offset) value
if the command went to the background and the replicate operation happened from the background
This cause the `wait` command to be ineffective.

By sending a simple write command (like `set`)(before the `wait` command). We make sure the
client `woff` value will be updated correctly and the `wait` command will be executed
as expected.

In addition this change make sure that connection that did not register the library will also
wait for the library to reach the replica (`wait` command waits for all the write that was performed by the
current connection).

This should fix failures such as: https://github.com/RedisGears/RedisGears/actions/runs/4870846502/jobs/8687088127

We should also consdiered whether or not this is a bug in Redis.