Add allow list to filter out globals.
Fix #902 add allow list and deny list to filter JS globals.

Any global which are not on the allow list will be removed.
If the global is also not appear of the deny list, we will output
a log message indicating we have a global which we do not considered
and we should check the porpose of this global and decide if we want to
filter it or not.

List of allow globals:

* Object
* Function
* Array
* Number
* parseFloat
* parseInt
* Infinity
* NaN
* undefined
* Boolean
* String
* Symbol
* Date
* Promise
* RegExp
* Error
* AggregateError
* RangeError
* ReferenceError
* SyntaxError
* TypeError
* URIError
* globalThis
* JSON
* Math
* Intl
* ArrayBuffer
* Uint8Array
* Int8Array
* Uint16Array
* Int16Array
* Uint32Array
* Int32Array
* Float32Array
* Float64Array
* Uint8ClampedArray
* BigUint64Array
* BigInt64Array
* DataView
* Map
* BigInt
* Set
* WeakMap
* WeakSet
* Proxy
* Reflect
* FinalizationRegistry
* WeakRef
* decodeURI
* decodeURIComponent
* encodeURI
* encodeURIComponent
* escape
* unescape
* isFinite
* isNaN
* console
* WebAssembly

List of deny globals:

* eval - Might be considered dangerous.
* EvalError - Sense we remove eval, this one is also not needed.
* SharedArrayBuffer - Needed for web workers which we are not supporting
* Atomics - Needed for web workers which we are not supporting

Require - https://github.com/RedisGears/v8-rs/pull/22