Added optional `onTriggerFired` argument to notification consumer API.
Fix #896.

The PR adds a new optional argument, `onTriggerFired`, that allows specify a function that will be invoke exactly when a notification happened and allows to read the data as it was at the time of the notification.

In addition, the PR refactor the API to use Redis post notification job where it is safe to perform write operations.

Example of the new API usage was added to the docs.