Refactor `FUNCTION LIST` code.
The regactoring rearrange the code that generate the output
for `FUNCTION LIST` command. The refactoring will enable us,
in the future, to return the information on other format (for
example on Redis `INFO` command). In addition, the refactoring
will enable (in the future) to aggregate the information from
all the shards using `LibMR`.