Refactor `FUNCTION LIST` code. (#912)
The refactoring rearrange the code that generate the output for `FUNCTION LIST` command. The refactoring will enable us, in the future, to return the information on other format (for example on Redis `INFO` command). In addition, the refactoring will enable (in the future) to aggregate the information from all the shards using `LibMR`.

Future Improvements:
1. Collect the `LibararyInfo` from all shards and aggregate it before returning to the user.
2. Return the information on Redis `info` command.