Combine library and function name to a single argument. (#939)
Combine library and function name to a single argument.

Fixes #920.

The PR combine the library and the function name to a single, `.` seperated argument
instead of 2 distinct arguments. The main goal is to make the API closer to Redis Function while not giving up on the library name which we believe is important.

The PR also forces the set of allowed char for functions and library name to be `[a-zA-Z0-9_][a-zA-Z0-9_]?` just like Redis Function.