Added support for `callAsync`.
Fix #937

The PR adds support for to call Redis commands that might block the client like [`blpop`](https://redis.io/commands/blpop/).
The new API added on this PR:

1. `client.callAsync` - run poterntially blocking Redis command, returns a promise object that will be resolved when the command finished.
2. `client.callAsyncRaw` - just like `client.callAsync` but do not perform utf8 decoding.

In order to support calling blocking commands, we use the new [`redismodule-rs`](https://github.com/RedisLabsModules/redismodule-rs) API, [`call_blocking`](https://github.com/RedisLabsModules/redismodule-rs/pull/333). The new API allows to call a blocking command and get a future CallReply which will be resolved when the command finished. This future CallReply must be aborted if the instance become replica. To supported that we maintain a new list per library that holds all the created future object (we call it `future_handlers`). In case the instance become replica, we can abort all future CallReplies and return an error to the v8 plugin which will raise the error on the user code (see `testCallAsyncBecomeReplica` test for demonstration).

Notice that the list hold a weak reference to the `future_handlers`, so if it was already resolved it will just be ignored. We use Redis cron job to clean those weak refernces.

From V8 plugin POV, when exeucting a blocking command, the V8 plugin will create a JS promise object that will be resolved by the callback that it set on the future CallReply.

Usage example:

```js
#!js api_version=1.0 name=lib

redis.registerAsyncFunction('my_blpop', async function(client, key, expected_val) {
    var res = null
    do {
        res = await client.block((c) => {
            return c.callAsync('blpop', key, '0');
        })
    } while (res[1] != expected_val);
    return res;
});
```

The following function will continue popping elements from the requested list up until it will encounter the requested value. In case the list is empty it will wait until elements will be added to the list.