Fixes according to recent librm changes (#974)
Recent changes of LibMR is lazy initialisation of threads and connections. So we can not wait for the connections to the other shards to be connected. Instead we can wait for it to be either connected or uninitialised (Which means that we got the topology and we will connect when needed).