Add Support for Handling Fatal Errors in Async Functions (#979)
The pull request (PR) introduces support for handling fatal errors in async functions. One of the primary challenges with async functions and fatal errors is that the callback set on the async function's future object will not be invoked if the async function encounters a fatal error, such as a timeout or an out-of-memory (OOM) condition. In such cases, the only way to resolve the promise is by registering a callback that will be triggered when the promise object is garbage collected. This is when an error can be returned to the user.

The PR also includes a comprehensive rearrangement of the JavaScript (JS) code invocation and promise handling. This is achieved through three key code refactoring steps:

1. Any invocation of JS code is performed using the `script_ctx` object, which utilizes the `call` and `run` functions. The `call` function is used for invoking JS functions, while `run` is used for running scripts. These functions handle the necessary tasks before and after executing JS code to ensure the proper functioning of timeout and OOM mechanisms.

2. Resolving or rejecting promises also goes through the `script_ctx` object, utilizing the `resolve` and `reject` functions. These functions handle the required tasks before and after executing JS code since resolving or rejecting a promise may involve the execution of JS code.

3. Setting promise handlers is achieved through a set of functions on the `script_ctx` object. These functions handle the checking of whether the promise has already been resolved or rejected. If so, the `on_done` closure is immediately called. Otherwise, the `on_done` closure is set as the resolver or rejecter for the promise object.

The third point primarily focuses on code organization and maintainability, reducing a significant amount of code duplication in promise handling.

The pull request (PR) also introduces two new debug commands:

1. `request_v8_gc_for_debugging`: This command forcibly triggers a JavaScript garbage collection (GC) for debugging purposes. It is intended to assist with debugging and analyzing memory-related issues.

2. `avoid_global_allow_list`: By using this command, the allow list for registering a function is bypassed. This is particularly useful when `--expose_gc` is used, which adds the `gc` global object that is not included in our allow list. Although the `gc` object is not allowed in our normal operations, it is necessary for debugging and testing purposes.

These new debug commands enhance the debugging capabilities and provide more flexibility for testing scenarios, allowing for better analysis and resolution of issues.