Set panick handler on v8 plugin (#984)
Set panic handler on v8 plugin. The plugin handler will call `RedisModule__Assert` to make sure we get a full crash report on panicking. Notice that we need to set it separately on the V8 plugin because the panic handler is enabled at the scope of an ELF and the V8 plugin is a separate `.so` file (separate ELF file).