Respect avoid replication traffic indication.
Fixes #988.

The PR adds support for avoid replication traffic.
If Redis requests to avoid replication traffic we should do the following:

1. Not allow locking Redis for Writes from within a async function
2. Stop processing stream record.

Notice that we do not stop processing key space notifications, this can be a violation of the atomic contruct.
The key space notification should stop because the clients should be paused in this situations.