Fix crash when resolving the same promise twice.
When resolving a promise we turing the resolver persisted value into a local value (`take_local` function).
This can only happened once, if happened more then once an assert will be raise.
When we resolved a promise more then once we tried to turn the persisted value to local value more then once
and raise the assert.

The fix is to wrap the persisted value with an Option and protect this function to be called once at the first
time and ignore it on the rest of the times.