Fix crash on benchmark when running with V8 main branch. (#996)
The crash started to happened since we merge this PR: https://github.com/RedisGears/RedisGears/pull/973

The purpose of the PR was to initialise the V8 backend lazy only when needed. A side effect was that the V8 might have been initialises on some other thread which are not the main thread. Unfortunately, the V8 platform must be initialised on the main thread (for some reason that I yet been able to understand) and this causes the crash.

The fix is to perform minimal initialisation steps on load time and leave the heavy V8 initialisation processes for the first time its been used. To achieve this, the PR adds a new `on_load` backend interface to `V8Backend` that will be called on loading time and allow the V8 backend to only initialise the platforms.