Build change to copy files to a second stage docker (#997)
* Fix crash on benchmark when running with V8 main branch.

The crash started to happened since we merge this PR: https://github.com/RedisGears/RedisGears/pull/973

The purpose of the PR was to initialise the V8 backend lazy only when needed.
A side effect was that the V8 might have been initialises on some other thread
which are not the main thread. Unfortunatly, the V8 palatform must be initialised
on the main thread (for some reason that I yet been able to understand) and this causes
the crash.

The fix is to perform minimal initialisation steps on load time and leave the heavy V8
initialisation processes for the first time its been used. To achieve this, the PR adds
a new `on_load` backend interaface to `V8Backend` that will be called on loading time
and allow the V8 backend to only initialise the platforms.

Depends on: https://github.com/RedisGears/v8-rs/pull/45

* run single failing tests with log output

* Fix failing tests

* Run all tests

* Move set_hook and global_ctx to on_load.

* checkrhel

* placeholder

* try

* restore

* sub with copy, to see size issue

* restoring

* pr comments

* restore

---------

Co-authored-by: meir <meir@redis.com>