Fix #1000, trim stream on post execution job. (#1001)
Fix #1000.

Trim the stream should be done on post execution job so we will make
sure the trimming happeneds last. otherwise, we might perform the
trimming on the key notification itself and this is not allow (we
can not perform any write operation inside a key space notification
callback, doing so can cause issues like replication reordering).