Fix #1005, avoid trigger key space notification inside function that trigger stream processing.
Fix #1005.

If a function trigger a stream processing by writting into a stream, when the stream processing
code were done it would have reset the `NotificationBlocker` and so it would trigger notifications
inside the function code.
To avoid that, the notification blocker will reset the value to the previous value that was set before.