Run stream trigger on post execution job.
Currently Stream trigger is invoke as part of the stream key space notification.
On of Redis constrains is that a module should not perform any write operation insdie
a key space notification callback. The PR moves the stream trigger invocation into
a post execution job so we will not violate the Redis constrains. Without this fix
we are getting a reordering of the replication order (the `xadd` command are replicated
last).