Apply loading timeout on AOF and Replication stream.
The loading timeout introduce on https://github.com/RedisGears/RedisGears/pull/999
should also be applied when loading data from AOF and when replica gets the
replication stream from its primary. To handle those cases, set the loading timeout
to apply also on `_rg_internals.function load` command (which is the command that
is written to AOF and sent on the replication stream).

In addition, the PR rename `rdb-lock-redis-timeout` to `loading-lock-redis-timeout`
so it will match also to AOF and not only RDB.

In addition, the PR makes sure that `_rg_internals.function load` is only invoked
on replica or during loading time.