Move the post notification job creation into `on_stream_touched`.
Today the post notification job is created after the stream data was read.
This is wrong because it means that we will read the data on the notification itself.
Moving the post notification job into `on_stream_touched` will make sure the entire
logic is happening inside the post notification job.