Fix connecting more than once, work-around a crash.
The crash is due to the library already existing (not having been
removed) the second time we want to debug it. The current solution uses
the "REPLACE" flag for the libraries to load, so that we don't care if
it existed, we just replace it all the time. This work-arounds the
crash, but doesn't fix anything, this will be done later.

Fixes the problem when it was impossible to connect for debugging more
than once. The problem was due to mistakenly closing the socket using
RAII. The current code fixes it by proper moving out of an enum variant,
instead of zeroing out the memory, what fixes the issue well.