Move  to deny_list (#1042)
On 8b8c91b we added the Iterator to the allow list. Its wrong to do it on a patch version because old 2.0 version might not be able to load RDB that was created by some newer 2.0.x version. It is better to put the Iterator on the deny list for patches versions on 2.0.