Add JavaScript debugging (#1032)
* js-debugging rework

* Step one

* work via cron

* Separate script compilation and evaluation

* Another (hundreds of) attempt to make it work.

* Add javascript example & allow for remote debugging in the deploy.js

* Working better

* Fix connecting more than once, work-around a crash.

The crash is due to the library already existing (not having been
removed) the second time we want to debug it. The current solution uses
the "REPLACE" flag for the libraries to load, so that we don't care if
it existed, we just replace it all the time. This work-arounds the
crash, but doesn't fix anything, this will be done later.

Fixes the problem when it was impossible to connect for debugging more
than once. The problem was due to mistakenly closing the socket using
RAII. The current code fixes it by proper moving out of an enum variant,
instead of zeroing out the memory, what fixes the issue well.

* Clean up the code a little and add the first test code.

* Update the code to support the going-to-be-merged inspector API.

Additionally:
1. Fixes a few warnings.
2. Disables the timeout check for the scripts being debugged.

* Adjust for the new code of v8-rs

* Remove unused function

* Remove unused imports

* Use the master branch of v8-rs

* Wait for the debugging connection to happen only within a time limit.

Helps to avoid having a server running while no client is attempting
to connect within some reasonable time period.

* Try the 1.70 as CI MSRV

* Address some review comments

* Add more comments about logging.

* Delete the unused DebuggerDeleteCallback type.

* Correct a doc-comment.

* Remove the unsafe Send implementation for the DebuggerServer

* Refactor function_load_with_args

* Don't allow for debugging when in cluster mode.

* Add more thorough debugging tests in Python

* Add more debugging tests.

* Fix the rebase

* Correct the JS example

* Address the review comments

* Check that a library is deleted after debugging