Fix nightly run failure (#1077)
Turns out it is not allowed to raise a termination exception on a co-routine, and as of v12.1 V8 asserts it. To fix the issue we catch the timeout error (termination exception) on `block` function and raise a regular none termination exception. This means that the user will now be able to catch timeout error on the co-routine code and handle it. Currently we do not force any timeout on co-routine so this change does not allow the user to do anything he could not yet already do.
successful nightly run: https://github.com/RedisGears/RedisGears/actions/runs/7491877476

(cherry picked from commit 0d1add66c46c4775e85c07295c66d83e1c48274a)